+++ test/integration/bigip/property/propertiesTftpProfile.js
@@ -46,7 +46,7 @@ describe('TFTP_Profile', function () {
 
         if (getProvisionedModules().includes('cgnat')) {
             properties.push({
-                name: 'logProfile',
+                name: 'algLogProfile',
                 inputValue: [undefined, { use: 'algLogProfile' }, undefined],
                 expectedValue: ['none', '/TEST_TFTP_Profile/Application/algLogProfile', 'none'],
                 extractFunction: (o) => ((typeof o.logProfile === 'object') ? o.logProfile.fullPath : o.logProfile),